+++ packages/widgetbook/lib/src/knobs/object_dropdown_knob.dart
@@ -19,6 +19,7 @@ class ObjectDropdownKnob<T> extends Knob<T?> {
     required this.options,
     super.description,
     this.labelBuilder,
+    super.defaultToNull,
   }) : super(isNullable: true);
 
   final List<T> options;