+++ packages/widgetbook/test/src/knobs/string_knob_test.dart
@@ -127,6 +127,34 @@ void main() {
           expect(find.textWidget(text), findsOneWidget);
         },
       );
+
+      testWidgets(
+        'given an initial value with defaultToNull=true, '
+        'then the value should initially be null with checkbox unchecked',
+        (tester) async {
+          const initialValue = 'Widgetbook';
+
+          await tester.pumpKnob(
+            (context) => Text(
+              context.knobs
+                  .stringOrNull(
+                    label: 'Knob',
+                    initialValue: initialValue,
+                    defaultToNull: true,
+                  )
+                  .toString(),
+            ),
+          );
+
+          expect(find.textWidget('null'), findsOneWidget);
+
+          final checkbox = tester.widget<Checkbox>(find.byType(Checkbox));
+          expect(checkbox.value, false);
+
+          await tester.findAndTap(find.byType(Checkbox));
+          expect(find.textWidget(initialValue), findsOneWidget);
+        },
+      );
     },
   );
 }