+++ packages/widgetbook/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- **FEAT**: Add `defaultToNull` parameter to nullable knobs to make them start in a `null` _(i.e. unchecked)_ state while having a non-null initial value. ([#1790](https://github.com/widgetbook/widgetbook/pull/1790) - by [@youpelegrace](https://github.com/youpelegrace))
+
 ## 3.22.0
 
 - **FIX**: Ensure slider label use max size ([#1823](https://github.com/widgetbook/widgetbook/pull/1823) - by [@EArminjon](https://github.com/EArminjon))